generate token and restrict permissions

+++ rest/api/views.py
@@ -1,6 +1,7 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 
 from rest.models import Blog
 from .serializers import BlogSerializer
@@ -8,6 +9,7 @@ from account.models import Account
 
 
 @api_view(['GET', ])
+@permission_classes((IsAuthenticated,))
 def api_detail_blog_view(request, slug):
 
     try:
@@ -21,12 +23,18 @@ def api_detail_blog_view(request, slug):
 
 
 @api_view(['PUT', ])
+@permission_classes((IsAuthenticated,))
 def api_update_blog_view(request, slug):
 
     try:
         blog = Blog.objects.get(slug=slug)
     except Blog.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+
+    if user != blog.author:
+        return Response({'response': "You don't have permission to edit that!"})
 
     if request.method == 'PUT':
         serializers_data = BlogSerializer(blog, data=request.data)
@@ -39,12 +47,18 @@ def api_update_blog_view(request, slug):
 
 
 @api_view(['DELETE', ])
+@permission_classes((IsAuthenticated,))
 def api_delete_blog_view(request, slug):
 
     try:
         blog = Blog.objects.get(slug=slug)
     except Blog.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+
+    if user != blog.author:
+        return Response({'response': "You don't have permission to edit that!"})
 
     if request.method == 'DELETE':
         operation = blog.delete()
@@ -57,8 +71,9 @@ def api_delete_blog_view(request, slug):
 
 
 @api_view(['POST', ])
+@permission_classes((IsAuthenticated,))
 def api_post_blog_view(request):
-    account = Account.objects.get(email='ucup@gmail.com')  # still harcode
+    account = request.user
     blog = Blog(author=account)
     if request.method == 'POST':
         serializers_data = BlogSerializer(blog, data=request.data)
